updated example CLI to use same controller code used in Lambda function

+++ backend/mainmodule/internal/services/fibonacci/fibonacci_response.go
@@ -1,9 +1,9 @@
 package fibonacci
 
 type FibonacciResponse struct {
-	Name     string  `json:"name"`
-	Time     float64 `json:"time"`
-	Error    string  `json:"error,omitempty"`
-	Max      int     `json:"max"`
-	Sequence []int   `json:"sequence"`
+	Name     string    `json:"name"`
+	Time     float64   `json:"time"`
+	Error    string    `json:"error,omitempty"`
+	Max      int       `json:"max"`
+	Sequence []float64 `json:"sequence"`
 }
